refactor(vue-integration): Moved all the plugin files to a src directory to make packaging more straightforward and seperating Cypress files from the plugin
