+++ integrations/cypress/cypress/plugins/index.ts
@@ -1,5 +1,5 @@
-import { encrypt } from '../support/utils/encrypt';
+import { encrypt } from '../../src/utils/encrypt';
 
 export default (on: Cypress.PluginEvents) => {
-  on('task', { encrypt });
+    on('task', { encrypt });
 };
