+++ integrations/cypress/src/commands/auth0-js/auth.ts
