+++ integrations/cypress/src/commands/auth0-js/index.ts
