+++ integrations/cypress/src/commands/general/destroy-simulation.ts
