+++ integrations/cypress/src/commands/general/index.ts
