+++ integrations/cypress/src/utils/spec.ts
@@ -4,8 +4,8 @@ import { createClient } from '@simulacrum/client';
 import type { TestState } from '../types';
 
 export interface MakeGetClientFromSpecOptions {
-  atom: Slice<TestState>;
-  port: number;
+    atom: Slice<TestState>;
+    port: number;
 }
